Move to action cable listener
