+++ enterprise/app/listeners/captain_listener.rb
@@ -10,23 +10,4 @@ class CaptainListener < BaseListener
     Captain::Llm::ContactNotesService.new(assistant, conversation).generate_and_update_notes if assistant.config['feature_memory'].present?
     Captain::Llm::ConversationFaqService.new(assistant, conversation).generate_and_deduplicate if assistant.config['feature_faq'].present?
   end
-
-  def copilot_message_created(event)
-    copilot_message = event.data[:copilot_message]
-    copilot_thread = copilot_message.copilot_thread
-    account = copilot_thread.account
-    user = copilot_thread.user
-
-    broadcast(account, [user.pubsub_token], COPILOT_MESSAGE_CREATED, copilot_message.push_event_data)
-  end
-
-  private
-
-  def broadcast(account, tokens, event_name, data)
-    return if tokens.blank?
-
-    payload = data.merge(account_id: account.id)
-
-    ::ActionCableBroadcastJob.perform_later(tokens.uniq, event_name, payload)
-  end
 end
